requirements.txt
groq==0.11.0
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.3.0
